pyproject.toml
[project]
name = "grpc-fastapi-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "redis>=7.2.0",
    "uvicorn>=0.41.0",
]
